TypeScript Conversion
