+++ app/(root)/page.tsx
@@ -2,4 +2,5 @@ import { redirect } from 'next/navigation';
 
 export default function Home() {
   redirect('/dashboard');
+  return <></>; 
 }
