+++ app/(root)/signup/page.tsx
@@ -1,20 +1,21 @@
 'use client';
 
-import React, { Suspense, useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useRouter, useSearchParams } from 'next/navigation';
+import { useRouter } from 'next/navigation';
+import { RootState, AppDispatch } from '@/app/redux/store';
 import {
     Box,
     Typography,
     TextField,
     Button,
     Link,
+    InputAdornment,
+    IconButton,
     CircularProgress,
     styled,
-    InputAdornment,
-    IconButton
 } from '@mui/material';
-import { performResetPassword } from '@/app/redux/actions/authActions';
+import { signUp } from '@/app/redux/actions/authActions';
 
 const MainBox = styled(Box)(({ theme }) => ({
     "& .loginLeft": {
@@ -117,7 +118,6 @@ const MainBox = styled(Box)(({ theme }) => ({
             fontWeight: "700",
             color: "#757575",
             fontSize: "15px",
-            fontWeight: "700",
             marginBottom: "20px",
             textAlign: "center",
             position: "relative",
@@ -127,7 +127,8 @@ const MainBox = styled(Box)(({ theme }) => ({
                 background: "#fff"
             },
             "&::before": {
-                background: "rgb(255,255,255)",
+                // can I delete this? it is duplicated 
+                // background: "rgb(255,255,255)",
                 background: "linear-gradient(90deg, rgba(255,255,255,1) 0%, rgba(224,224,224,1) 15%, rgba(255,255,255,1) 50%, rgba(224,224,224,1) 85%, rgba(255,255,255,1) 100%)",
                 width: "100%",
                 height: "1px",
@@ -171,53 +172,53 @@ const MainBox = styled(Box)(({ theme }) => ({
     }
 }));
 
-export default function ResetPasswordPageWrapper() {
-    return (
-        <Suspense fallback={<div>Loading...</div>}>
-            <RestPasswordPage />
-        </Suspense>
-    );
-}
-
-function RestPasswordPage() {
-    const [newPassword, setNewPassword] = useState('');
-    const [confirmPassword, setConfirmPassword] = useState('');
-    const [errorMessage, setErrorMessage] = useState('');
-    const [showNewPassword, setShowNewPassword] = useState(false);
-    const [showConfirmPassword, setShowConfirmPassword] = useState(false);
+export default function SingupPage() {
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
+    const [firstName, setFirstName] = useState('');
+    const [lastName, setLastName] = useState('');
+    const [mobile, setMobile] = useState('');
+    const dispatch = useDispatch<AppDispatch>();
+    const { loading, user, signupData, error } = useSelector((state: RootState) => state.user);
     const router = useRouter();
-    const dispatch = useDispatch();
-    const { loading, error, resetPasswordMessage } = useSelector((state) => state.user);
-    const searchParams = useSearchParams();
-    const username = searchParams.get('username');
-    const code = searchParams.get('code');
 
-    const handleResetPassword = (e) => {
+    const [showPassword, setShowPassword] = React.useState(false);
+
+    const handleSignup = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
-        if (newPassword !== confirmPassword) {
-            setErrorMessage('Passwords do not match');
+        if (!firstName || !lastName || !email || !password || !mobile) {
             return;
         }
-        dispatch(performResetPassword({
-            'Agentlang.Kernel.Identity/ConfirmForgotPassword': {
-                Password: newPassword,
-                Username: username,
-                ConfirmationCode: code
+        dispatch(signUp({
+            'Agentlang.Kernel.Identity/SignUp': {
+                'User': {
+                    'Agentlang.Kernel.Identity/User': {
+                        Name: `${firstName} ${lastName}`,
+                        FirstName: firstName,
+                        LastName: lastName,
+                        Email: email,
+                        UserData: {
+                            PhoneNumber: mobile
+                        },
+                        Password: password
+                    }
+                }
             }
-        })).then((response) => {
-            console.log('Password reset successful:', response);
-        });
-
-        setNewPassword('');
-        setConfirmPassword('');
-        setErrorMessage('');
+        }, email));
     };
 
-    const handleToggleNewPassword = () => setShowNewPassword((prev) => !prev);
-    const handleToggleConfirmPassword = () => setShowConfirmPassword((prev) => !prev);
+    useEffect(() => {
+        if (user) {
+            router.push('/dashboard');
+        }
+    }, [user, router]);
 
-    if (resetPasswordMessage && !error) {
-        router.push('/login');
+    const handleTogglePassword = () => {
+        setShowPassword((prev) => !prev);
+    };
+
+    if (signupData && !error) {
+        router.push('/signup-otp');
     }
 
     return (
@@ -235,63 +236,78 @@ function RestPasswordPage() {
                 <Box className='loginRight'>
                     <Box className='formBox'>
                         <Typography variant="h4">
-                            Reset Password
+                            Create an Account
                         </Typography>
-                        <Typography className='subHeadingText' whiteSpace={'nowrap'}>
-                            Your new password must be different from previous one
+                        <Typography className='subHeadingText'>
+                            Please enter your details
                         </Typography>
                         <Box
                             component="form"
-                            onSubmit={handleResetPassword}
+                            onSubmit={handleSignup}
                         >
                             <TextField
                                 className='textField'
+                                id="outlined-basic"
+                                placeholder="First Name"
                                 variant="outlined"
-                                placeholder="New Password"
-                                type={showNewPassword ? "text" : "password"}
-                                fullWidth
-                                value={newPassword}
-                                onChange={(e) => setNewPassword(e.target.value)}
-                                InputProps={{
-                                    endAdornment: (
-                                        <InputAdornment position="end">
-                                            <IconButton onClick={handleToggleNewPassword} edge="end">
-                                                {showNewPassword ? (
-                                                    <img src={"/images/icons/eye-on.svg"} alt='eye-on' />
-                                                ) : (
-                                                    <img src={"/images/icons/eye-off.svg"} alt='eye-off' />
-                                                )}
-                                            </IconButton>
-                                        </InputAdornment>
-                                    ),
-                                }}
+                                value={firstName}
+                                onChange={(e) => setFirstName(e.target.value)}
                             />
-
+                            <TextField
+                                className='textField'
+                                id="outlined-basic"
+                                placeholder="Last Name"
+                                variant="outlined"
+                                value={lastName}
+                                onChange={(e) => setLastName(e.target.value)}
+                            />
+                            <TextField
+                                className='textField'
+                                id="outlined-basic"
+                                placeholder="Mobile Number"
+                                variant="outlined"
+                                type='tel'
+                                value={mobile}
+                                onChange={(e) => setMobile(e.target.value)}
+                            />
+                            <TextField
+                                className='textField'
+                                id="outlined-basic"
+                                placeholder="Email Id"
+                                InputLabelProps={{
+                                    shrink: false
+                                }}
+                                variant="outlined"
+                                value={email}
+                                onChange={(e) => setEmail(e.target.value)}
+                            />
                             <TextField
                                 className='textField'
                                 variant="outlined"
-                                placeholder="Confirm Password"
-                                type={showConfirmPassword ? "text" : "password"}
+                                placeholder="Password"
+                                type={showPassword ? "text" : "password"}
+                                InputLabelProps={{
+                                    shrink: false
+                                }}
                                 fullWidth
-                                value={confirmPassword}
-                                onChange={(e) => setConfirmPassword(e.target.value)}
+                                value={password}
+                                onChange={(e) => setPassword(e.target.value)}
                                 InputProps={{
                                     endAdornment: (
                                         <InputAdornment position="end">
-                                            <IconButton onClick={handleToggleConfirmPassword} edge="end">
-                                                {showConfirmPassword ? (
-                                                    <img src={"/images/icons/eye-on.svg"} alt='eye-on' />
-                                                ) : (
-                                                    <img src={"/images/icons/eye-off.svg"} alt='eye-off' />
-                                                )}
+                                            <IconButton onClick={handleTogglePassword} edge="end">
+                                                {showPassword ? <img src={"/images/icons/eye-on.svg"} alt='eye-on' /> : <img src={"/images/icons/eye-off.svg"} alt='eye-off' />}
                                             </IconButton>
                                         </InputAdornment>
                                     ),
                                 }}
                             />
-                            {errorMessage && (
-                                <Typography variant="body2" color="error" sx={{ mb: 2 }}>
-                                    {errorMessage}
+                            {error && (
+                                <Typography
+                                    variant="body2"
+                                    color="error"
+                                >
+                                    {error}
                                 </Typography>
                             )}
                             <Button
@@ -303,31 +319,18 @@ function RestPasswordPage() {
                                 sx={{ mt: 2 }}
                                 className='signInButton'
                             >
-                                {loading ? <CircularProgress size={24} /> : 'Reset Password'}
+                                {loading ? <CircularProgress size={24} /> : 'Sign up'}
                             </Button>
                         </Box>
-
+                        <Typography className='noAccount'>
+                            Already have an account?{' '}
+                            <Link href="/login" underline="hover" color="primary">
+                                Sign in
+                            </Link>
+                        </Typography>
                     </Box>
                 </Box>
             </Box>
-            {resetPasswordMessage && !error && (
-                <Typography
-                    variant="body2"
-                    color="success"
-                    sx={{ position: 'absolute', bottom: '150px', left: '50%', transform: 'translateX(-50%)' }}
-                >
-                    Password reset successfully! You can now log in with your new password.
-                </Typography>
-            )}
-            {error && (
-                <Typography
-                    variant="body2"
-                    color="error"
-                    sx={{ position: 'absolute', bottom: '150px', left: '50%', transform: 'translateX(-50%)' }}
-                >
-                    {error}
-                </Typography>
-            )}
         </MainBox>
     );
 }
