+++ app/components/Input/StyledInput.tsx
@@ -1,0 +1,128 @@
+import { FormHelperText, styled, TextField, Box } from '@mui/material';
+import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
+import { ReactNode } from 'react';
+
+interface CustomInputProps {
+  width?: string;
+  margin?: string;
+  padding?: string;
+  height?: string;
+  error?: boolean;
+}
+
+export const StyledInput = styled(TextField, {
+  shouldForwardProp: (prop) =>
+    !['width', 'margin', 'padding', 'height', 'error'].includes(prop as string),
+})<CustomInputProps>(({ theme, width, margin, padding, height, error }) => ({
+  '& .MuiOutlinedInput-root': {
+    borderRadius: '4px',
+    height: height || '36px',
+    backgroundColor: '#FFF',
+    fontSize: '12px',
+    '& fieldset': {
+      borderColor: '#D6DCE1',
+    },
+    '&.Mui-disabled': {
+      backgroundColor: '#E5E7EB !important',
+      color: '#6B7280',
+    },
+  },
+  '& fieldset': {
+    borderColor: error ? theme.palette.error.main : '#D6DCE1',
+  },
+  '&.bold-input .MuiInputBase-input': {
+    fontWeight: 'bold',
+    fontSize: '16px',
+  },
+  '&:hover fieldset': {
+    borderColor: error ? theme.palette.error.main : '#D6DCE1',
+  },
+  '&.Mui-focused fieldset': {
+    borderColor: error ? theme.palette.error.main : '#D6DCE1',
+  },
+  '& input::-webkit-outer-spin-button, & input::-webkit-inner-spin-button': {
+    WebkitAppearance: 'none',
+    margin: 0,
+  },
+  '& input[type="number"]': {
+    MozAppearance: 'textfield',
+  },
+  '& .MuiFormHelperText-root': {
+    margin: '4px 0 0 0',
+    fontSize: '0.75rem',
+    lineHeight: '12px',
+    color: error ? theme.palette.error.main : '#6A7178',
+    '&.Mui-error': {
+      color: theme.palette.error.main,
+    },
+  },
+  width: width || '100%',
+  margin: margin || '0',
+  padding: padding || '0',
+}));
+
+export const StyledCommentInput = styled(TextField, {
+  shouldForwardProp: (prop) =>
+    !['width', 'margin', 'padding', 'height'].includes(prop as string),
+})<CustomInputProps>(({ width, margin, padding, height }) => ({
+  '& .MuiOutlinedInput-root': {
+    borderRadius: '4px',
+    height: height || '88px',
+    fill: '#FDFDFD',
+    strokeWidth: '1px',
+    stroke: '#D6DCE1',
+    boxShadow: '0px 2px 4px 0px #EBEBEB inset',
+    backgroundColor: '#FFFFFFF',
+    fontSize: '12px',
+    '& fieldset': {
+      borderColor: '#D6DCE1',
+    },
+  },
+  width: width || '100%',
+  margin: margin || '0',
+  padding: padding || '0',
+}));
+
+export const StyledFormHelperText = styled(FormHelperText)(({ theme }) => ({
+  color: theme.palette.info.main,
+  marginLeft: 0,
+  fontSize: '0.75rem',
+}));
+
+interface StyledFormInfoTextProps {
+  children: ReactNode;
+}
+
+export const StyledFormInfoText = styled(({ children, ...props }: StyledFormInfoTextProps) => (
+  <Box sx={{ display: 'flex', gap: 0, lineHeight: 1 }}>
+    <InfoOutlinedIcon
+      sx={{
+        fontSize: '0.75rem',
+        color: '#757575',
+        display: 'flex',
+        marginRight: '2px',
+      }}
+    />
+    <FormHelperText
+      sx={{
+        display: 'inline-flex',
+        margin: '0 !important',
+        lineHeight: '1 !important',
+        padding: 0,
+      }}
+      {...props}
+    >
+      {children}
+    </FormHelperText>
+  </Box>
+))(() => ({
+  color: '#757575',
+  marginLeft: 0,
+  fontSize: '0.75rem',
+  display: 'inline-flex',
+  alignItems: 'center',
+  '& .MuiFormHelperText-root': {
+    margin: 0,
+    display: 'inline',
+  },
+}));
