--- app/components/Projects/Table/ProjectTable.jsx
@@ -1,102 +1,0 @@
-import { ColumnManagementStyles, FilterPanelStyles, StyledDataGrid } from "../../AllocationTable/styles/StyledDataGrid";
-import {
-    GridColumnMenu,
-    useGridApiRef
-  } from '@mui/x-data-grid-premium';
-import ProjectToolbar from "../../Toolbar/ProjectToolbar";
-import { useState } from "react";
-function CustomColumnMenu(props) {
-    return (
-      <GridColumnMenu
-        {...props}
-        slots={{
-            columnMenuAggregationItem: null,
-            columnMenuGroupingItem: null,
-        }}
-      />
-    );
-  }
-  
-const ProjectTable = ({columns, rows, loading, apiRef}) => {
-    const [filterButtonEl, setFilterButtonEl] = useState(null);
-    return (
-        <StyledDataGrid
-            apiRef={apiRef}
-            columns={columns}
-            rows={rows}
-            hideFooter={true}
-            loading={loading}
-            initialState={{
-                sorting: {
-                  sortModel: [
-                    { field: 'Name', sort: 'asc' },
-                  ],
-                },
-                columns: {
-                  columnVisibilityModel: {
-                    ProjectSponsor: false, 
-                    ProjectManager: false, 
-                    Location: false,
-                    Budget: false,
-                    BudgetCurrency:false,
-                  },
-                },
-              }}
-            slots={{
-                toolbar: ProjectToolbar,
-                columnMenu: CustomColumnMenu
-            }}
-            sx={{
-                height: '95vh',
-                '& .MuiDataGrid-columnHeader': {
-                    padding: '0 16px',
-                    borderRight: 'none',
-                }
-            }}
-            slotProps={{
-                loadingOverlay: {
-                variant: 'skeleton',
-                noRowsVariant: 'skeleton',
-                },
-                panel: {
-                anchorEl: filterButtonEl,
-                className: 'parent-grid-panel',
-                },
-                toolbar: {
-                    setFilterButtonEl,
-                },
-                columnsPanel: {
-                className: 'styleColumnMenu',
-                sx: ColumnManagementStyles,
-                },
-                filterPanel: {
-                columnsSort: 'asc',
-                className: 'filterPopup',
-                filterFormProps: {
-                    columnInputProps: {
-                    size: 'small',
-                    sx: { mt: 'auto' },
-                    },
-                    operatorInputProps: {
-                    size: 'small',
-                    sx: { mt: 'auto' },
-                    },
-                    valueInputProps: {
-                    InputComponentProps: {
-                        size: 'small',
-                    },
-                    },
-                    deleteIconProps: {
-                    sx: {
-                        '& .MuiSvgIcon-root': { color: '#d32f2f' },
-                    },
-                    },
-                },
-                sx: FilterPanelStyles,
-                },
-            }}
-        />
-    );
-}
-
-export default ProjectTable;
