+++ app/components/ResourceAllocation/component/OrganizationAllocation.tsx
@@ -1,7 +1,9 @@
 'use client';
+import { Box } from '@mui/material';
 import AllocationGrid from '@/app/components/AllocationTable/AllocationGrid';
+import type { GridColDef } from '@mui/x-data-grid-pro';
 
-const organizationColumnConfig = [
+const organizationColumnConfig: GridColDef[] = [
   { field: 'teams', headerName: 'Organizaion Name', width: 240 },
   {
     field: 'project',
@@ -24,6 +26,15 @@ export default function OrganizationAllocation() {
         <AllocationGrid
           groupBy="organization"
           columns={organizationColumnConfig}
+          loading={false}
+          selectedTeam={null}
+          setSelectedTeam={() => {}}
+          initialState={{}}
+          startDate={new Date()}
+          endDate={new Date()}
+          toolbarComponent={null}
+          NoRowsOverlay={null}
+          mode="view"
         />
       </Box>
     </>
