+++ app/components/TableIcons/AllTeamsIcon.tsx
@@ -1,7 +1,15 @@
 import React from 'react';
-import { SvgIcon } from '@mui/material';
+import { SvgIcon, SvgIconProps } from '@mui/material';
 
-const AllTeamsIcon = ({ color = '#344665', fontSize = 'medium', ...props }) => {
+interface AllTeamsIconProps extends SvgIconProps {
+  customColor?: string;
+}
+
+const AllTeamsIcon: React.FC<AllTeamsIconProps> = ({
+  customColor = '#344665',
+  fontSize = 'medium',
+  ...props
+}) => {
   return (
     <SvgIcon {...props} fontSize={fontSize} viewBox="0 0 22 16">
       <defs>
@@ -12,12 +20,12 @@ const AllTeamsIcon = ({ color = '#344665', fontSize = 'medium', ...props }) => {
       <g clipPath="url(#clip0_7_17)">
         <path
           d="M5.36268 7.59156L3.174 7.59142H3.17396C2.0838 7.59142 1.20005 8.47517 1.20005 9.56533V10.6523C1.20005 10.7627 1.11051 10.8523 1.00005 10.8523C0.889592 10.8523 0.800049 10.7627 0.800049 10.6523V9.56533C0.800049 8.25425 1.86289 7.19142 3.17396 7.19142H3.17402L5.19782 7.19116C5.24784 7.32714 5.30287 7.46069 5.36268 7.59156ZM6.64551 1.0553C6.54597 1.14907 6.4498 1.24635 6.35721 1.34696C6.062 1.17104 5.71658 1.06968 5.34788 1.06968C4.25771 1.06968 3.37396 1.95343 3.37396 3.04359C3.37396 3.96314 4.00274 4.73583 4.8538 4.95516C4.84986 5.04215 4.84788 5.12961 4.84788 5.2175C4.84788 5.26686 4.8485 5.31608 4.84975 5.36515C3.77783 5.13627 2.97396 4.18381 2.97396 3.04359C2.97396 1.73251 4.0368 0.669678 5.34788 0.669678C5.82692 0.669678 6.27247 0.81124 6.64551 1.0553ZM18.8261 7.19142H18.8261C20.1372 7.19142 21.2 8.25425 21.2 9.56533V10.6523C21.2 10.7627 21.1105 10.8523 21 10.8523C20.8896 10.8523 20.8 10.7627 20.8 10.6523V9.56533C20.8 8.47517 19.9163 7.59142 18.8261 7.59142L15.7679 7.59158C15.8277 7.46069 15.8827 7.32714 15.9327 7.19115L18.8261 7.19142ZM15.1348 1.78115C15.0533 1.67266 14.9679 1.56723 14.879 1.46503C15.3144 0.976411 15.9476 0.669678 16.6522 0.669678C17.9633 0.669678 19.0261 1.73251 19.0261 3.04359C19.0261 4.35467 17.9633 5.4175 16.6522 5.4175C16.5254 5.4175 16.4012 5.40761 16.2801 5.38858C16.2818 5.33166 16.2827 5.27463 16.2827 5.2175C16.2827 5.13863 16.2811 5.0601 16.2779 4.98196C16.3992 5.00529 16.5243 5.0175 16.6522 5.0175C17.7424 5.0175 18.6261 4.13375 18.6261 3.04359C18.6261 1.95343 17.7424 1.06968 16.6522 1.06968C16.042 1.06968 15.4968 1.34637 15.1348 1.78115Z"
-          fill={color}
-          stroke={color}
+          fill={customColor}
+          stroke={customColor}
         />
         <path
           d="M15.7827 14.3479V13.0435C15.7827 11.6028 14.6147 10.4348 13.174 10.4348H7.9566C6.51585 10.4348 5.3479 11.6028 5.3479 13.0435V14.3479"
-          stroke={color}
+          stroke={customColor}
           strokeWidth="1.4"
           strokeLinecap="round"
           strokeLinejoin="round"
@@ -25,7 +33,7 @@ const AllTeamsIcon = ({ color = '#344665', fontSize = 'medium', ...props }) => {
         />
         <path
           d="M10.5652 7.82628C12.006 7.82628 13.1739 6.65833 13.1739 5.21758C13.1739 3.77684 12.006 2.60889 10.5652 2.60889C9.1245 2.60889 7.95654 3.77684 7.95654 5.21758C7.95654 6.65833 9.1245 7.82628 10.5652 7.82628Z"
-          stroke={color}
+          stroke={customColor}
           strokeWidth="1.4"
           strokeLinecap="round"
           strokeLinejoin="round"
