+++ app/utils/authUtils.ts
@@ -1,5 +1,5 @@
 // Save access token to localStorage
-export const saveToken = (token) => {
+export const saveToken = (token: string): null => {
   if (typeof window !== "undefined" && window.localStorage) {
     localStorage.setItem('token', token);
   }
@@ -7,7 +7,7 @@ export const saveToken = (token) => {
 };
 
 // Get access token from localStorage
-export const getToken = () => {
+export const getToken = (): string | null => {
   if (typeof window !== "undefined" && window.localStorage) {
     return localStorage.getItem("token");
   }
@@ -15,13 +15,13 @@ export const getToken = () => {
 };
 
 // Clear access token from localStorage
-export const clearToken = () => {
+export const clearToken = (): void => {
   localStorage.removeItem('token');
   localStorage.removeItem('refreshToken');
 };
 
 // Save refresh token to localStorage
-export const saveRefreshToken = (refreshToken) => {
+export const saveRefreshToken = (refreshToken: string): null => {
   if (typeof window !== "undefined" && window.localStorage) {
     localStorage.setItem('refreshToken', refreshToken);
   }
@@ -29,23 +29,29 @@ export const saveRefreshToken = (refreshToken) => {
 };
 
 // Get refresh token from localStorage
-export const getRefreshToken = () => {
+export const getRefreshToken = (): string | null => {
   return localStorage.getItem('refreshToken');
 };
 
 // Clear refresh token from localStorage
-export const clearRefreshToken = () => {
+export const clearRefreshToken = (): void => {
   localStorage.removeItem('refreshToken');
 };
 
 // Clear all authentication data (logout helper)
-export const clearAuth = () => {
+export const clearAuth = (): void => {
   clearToken();
   clearRefreshToken();
 };
 
 // Check user roles (RBAC utility)
-export const hasRole = (user, requiredRoles) => {
+
+// not sure what the uswr interface is
+interface User {
+  roles: string[];
+}
+
+export const hasRole = (user: User | null | undefined, requiredRoles: string []): boolean => {
   if (!user || !requiredRoles) return false;
   return requiredRoles.some((role) => user.roles.includes(role));
 };
