+++ package.json
@@ -47,6 +47,7 @@
     "@types/pg": "^8.15.2",
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.0",
+    "@types/react-grid-layout": "^1.3.5",
     "@types/react-redux": "^7.1.34",
     "@types/yup": "^0.29.14",
     "concurrently": "^9.1.2",
